django/code/requirements.txt
django
Pillow
django-extensions
requests
djangorestframework